Splitting the VM tests out into Managed and Unmanaged for manageability
